+++ client/apps/game/src/ui/components/resources/travel-info.tsx
@@ -1,13 +1,14 @@
 import { GRAMS_PER_KG } from "@/ui/constants";
 import { ResourceIcon } from "@/ui/elements/resource-icon";
+import { calculateDonkeysNeeded, currencyFormat, divideByPrecision, multiplyByPrecision } from "@/ui/utils/utils";
 import {
-  calculateDonkeysNeeded,
-  currencyFormat,
-  divideByPrecision,
+  configManager,
+  getBalance,
   getTotalResourceWeight,
-  multiplyByPrecision,
-} from "@/ui/utils/utils";
-import { configManager, getBalance, ResourcesIds, type ID, type Resource } from "@bibliothecadao/eternum";
+  ResourcesIds,
+  type ID,
+  type Resource,
+} from "@bibliothecadao/eternum";
 import { useDojo, useUIStore } from "@bibliothecadao/react";
 import { useEffect, useMemo, useState } from "react";
