+++ client/sdk/packages/core/src/utils/resources.ts
@@ -1,10 +1,10 @@
 import { getComponentValue } from "@dojoengine/recs";
 import { getEntityIdFromKeys } from "@dojoengine/utils";
 import { getStartingResources } from ".";
-import { resources, ResourcesIds } from "../constants";
+import { BuildingType, resources, ResourcesIds } from "../constants";
 import { ClientComponents } from "../dojo";
 import { ResourceManager } from "../modelManager";
-import { configManager } from "../modelManager/ConfigManager";
+import { ClientConfigManager, configManager } from "../modelManager/ConfigManager";
 import { ID, Resource } from "../types";
 import { unpackResources } from "./packed-data";
 
@@ -65,4 +65,25 @@ export const getQuestResources = (realmEntityId: ID, components: ClientComponent
   const realm = getComponentValue(components.Realm, getEntityIdFromKeys([BigInt(realmEntityId)]));
   const resourcesProduced = realm ? unpackResources(realm.produced_resources) : [];
   return getStartingResources(resourcesProduced);
+};
+
+export const getTotalResourceWeight = (resources: Array<Resource | undefined>) => {
+  const configManager = ClientConfigManager.instance();
+
+  return resources.reduce(
+    (total, resource) =>
+      total + (resource ? resource.amount * configManager.getResourceWeight(resource.resourceId) || 0 : 0),
+    0,
+  );
+};
+
+export const isResourceProductionBuilding = (buildingId: BuildingType) => {
+  return (
+    buildingId === BuildingType.Resource ||
+    buildingId === BuildingType.Farm ||
+    buildingId === BuildingType.FishingVillage ||
+    buildingId === BuildingType.Barracks ||
+    buildingId === BuildingType.ArcheryRange ||
+    buildingId === BuildingType.Stable
+  );
 };
